# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Red Intelfon - Workflow de Activaciones API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /procesos/{id}/etapas:
    get:
      operationId: listEtapasProceso
      tags: [etapas]
      summary: Listar etapas de un proceso
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de etapas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EtapaProceso"
  /procesos/{id}/etapas/{numeroEtapa}:
    get:
      operationId: getEtapa
      tags: [etapas]
      summary: Obtener una etapa específica
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: numeroEtapa
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Etapa encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EtapaDetalle"
  /procesos/{id}/etapas/{numeroEtapa}/completar:
    post:
      operationId: completarEtapa
      tags: [etapas]
      summary: Marcar etapa como completada
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: numeroEtapa
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Etapa completada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EtapaProceso"
        "400":
          description: Checklist incompleto
  /procesos/{id}/etapas/{numeroEtapa}/justificar:
    post:
      operationId: justificarEtapa
      tags: [etapas]
      summary: Justificar retraso de SLA
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: numeroEtapa
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JustificacionInput"
      responses:
        "200":
          description: Justificación registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EtapaProceso"

  # Checklist
...
  /notificaciones/{id}/leer:
    put:
      operationId: marcarNotificacionLeida
      tags: [notificaciones]
      summary: Marcar notificación como leída
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notificación marcada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notificacion"
...
components:
  schemas:
...
    Proceso:
      type: object
      required: [id, numeroPreoferta, estadoActual, clienteNombre, prioridad, fijado, fechaInicio, slaGlobalHoras]
      properties:
        id:
          type: integer
        numeroPreoferta:
          type: string
        estadoActual:
          type: string
          enum: [en_espera, en_fase_1, en_fase_2, en_fase_3, en_fase_4, en_fase_5, completado]
        clienteNombre:
          type: string
        clienteEmail:
          type: ["string", "null"]
        clienteTelefono:
          type: ["string", "null"]
        fechaInicio:
          type: string
          format: date-time
        fechaFinEstimada:
          type: ["string", "null"]
          format: date-time
        fechaFinReal:
          type: ["string", "null"]
          format: date-time
        slaGlobalHoras:
          type: integer
        prioridad:
          type: string
          enum: [baja, media, alta, urgente]
        fijado:
          type: boolean
        usuarioCreadorId:
          type: ["integer", "null"]
        etapaActual:
          type: ["integer", "null"]
        slaVencido:
          type: boolean
...
    EtapaProceso:
      type: object
      required: [id, idProceso, numeroEtapa, estado, slaEtapaHoras]
      properties:
        id:
          type: integer
        idProceso:
          type: integer
        numeroEtapa:
          type: integer
        estado:
          type: string
          enum: [pendiente, activa, completada]
        fechaInicio:
          type: ["string", "null"]
          format: date-time
        fechaFin:
          type: ["string", "null"]
          format: date-time
        slaEtapaHoras:
          type: integer
        usuarioResponsableId:
          type: ["integer", "null"]
        checklistCompletado:
          type: boolean
        nombreEtapa:
          type: string
        slaVencido:
          type: boolean
        minutosRestantes:
          type: ["integer", "null"]
    EtapaDetalle:
      type: object
      required: [id, idProceso, numeroEtapa, estado, slaEtapaHoras, checklist]
      properties:
        id:
          type: integer
        idProceso:
          type: integer
        numeroEtapa:
          type: integer
        estado:
          type: string
        fechaInicio:
          type: ["string", "null"]
          format: date-time
        fechaFin:
          type: ["string", "null"]
          format: date-time
        slaEtapaHoras:
          type: integer
        usuarioResponsableId:
          type: ["integer", "null"]
        checklistCompletado:
          type: boolean
        nombreEtapa:
          type: string
        descripcionEtapa:
          type: ["string", "null"]
        areasInvolucradas:
          type: array
          items:
            type: string
        slaVencido:
          type: boolean
        minutosRestantes:
          type: ["integer", "null"]
        checklist:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        proceso:
          $ref: "#/components/schemas/Proceso"
    JustificacionInput:
      type: object
      required: [justificacion]
      properties:
        justificacion:
          type: string

    # Checklist
    ChecklistItem:
      type: object
      required: [id, idEtapaProceso, descripcion, completado]
      properties:
        id:
          type: integer
        idEtapaProceso:
          type: integer
        descripcion:
          type: string
        completado:
          type: boolean
        notas:
          type: ["string", "null"]
        fechaCompletado:
          type: ["string", "null"]
          format: date-time
        usuarioQuienCompletoId:
          type: ["integer", "null"]
...
    Notificacion:
      type: object
      required: [id, usuarioDestinoId, tipo, titulo, mensaje, leido, fechaCreacion]
      properties:
        id:
          type: integer
        usuarioDestinoId:
          type: integer
        idProceso:
          type: ["integer", "null"]
        tipo:
          type: string
          enum: [etapa_lista, sla_proximo, sla_vencido, chat_nuevo, comentario_nuevo, documento_faltante]
        titulo:
          type: string
        mensaje:
          type: string
        leido:
          type: boolean
        fechaCreacion:
          type: string
          format: date-time

    # Admin - Config Etapas
...