# lib/api-zod/src/generated/types/etapaProceso.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Red Intelfon - Workflow de Activaciones API
 * OpenAPI spec version: 0.1.0
 */
import type { EtapaProcesoEstado } from './etapaProcesoEstado';

export interface EtapaProceso {
  id: number;
  idProceso: number;
  numeroEtapa: number;
  estado: EtapaProcesoEstado;
  /** @nullable */
  fechaInicio?: Date | null;
  /** @nullable */
  fechaFin?: Date | null;
  slaEtapaHoras: number;
  /** @nullable */
  usuarioResponsableId?: number | null;
  checklistCompletado?: boolean;
  nombreEtapa?: string;
  slaVencido?: boolean;
  /** @nullable */
  minutosRestantes?: number | null;
}